Adding fileutils.DevNull convenience
The intended use case is with os.Exec.Cmd - it is often convenient to
intitialize a cmd struct in the following way:

```go
&exec.Cmd{
  //...
  Stdout: os.Stdout,
  Stderr: os.Stderr,
}
```

However, an attempt to use `os.DevNull` does not work the same way, as
that is actually just the string `"/dev/null"`.  Adding these io.Writer
handles to `fileutils` will help remove some confusion by making all
three accessible in the same package.  Also adding os.Stdin for the sake
of completion